src/main/java/ru/job4j/collection/SimpleQueue.java
package ru.job4j.collection;

import java.util.Stack;

public class SimpleQueue<T> {

    private final Stack<T> in = new Stack<>();
    private final Stack<T> out = new Stack<>();

    public T poll() {
        if (out.empty()) {
            while (!in.empty()) {
                out.push(in.pop());
            }
        }
        return out.pop();
    }

    public void push(T value) {
        in.push(value);
    }

    public static void main(String[] args) {
        SimpleQueue<Integer> queue = new SimpleQueue<>();
        queue.push(1);
        queue.push(2);
        queue.push(3);
        System.out.println(queue.poll());
        System.out.println(queue.poll());
        System.out.println(queue.poll());
    }
}
